+++ synapse-edison/src/test/java/de/otto/synapse/edison/health/FixedChannelHealthIndicatorTest.java
@@ -1,0 +1,14 @@
+package de.otto.synapse.edison.health;
+
+import java.util.HashSet;
+import java.util.List;
+
+
+public class FixedChannelHealthIndicatorTest extends AbstractChannelHealthIndicatorTest{
+
+
+    @Override
+    protected AbstractChannelHealthIndicator createHealthIndicator(List<String> channelNames) {
+        return new FixedChannelHealthIndicator(new HashSet<>(channelNames));
+    }
+}