+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Adds new NitriteStateRepository, including support for secondary indexes.
 * Introduces MessageSender interface
 * Adds a simple TeeMessageSender that is sending messages to multiple delegate MessageSenders.
+* Introduces FixedChannelHealthIndicator to measure health of asynchronously added eventsources 
 
 ## 0.20.0
 *synapse-core*